% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renalFunction.R
\name{adultCreatinine}
\alias{adultCreatinine}
\title{Assumed serum creatinine of an adult with normal kidneys}
\usage{
adultCreatinine(sex)
}
\arguments{
\item{sex}{\code{"male"} or \code{"female"}}
}
\value{
serum creatinine in mg/dL
}
\description{
Assumed serum creatinine of an adult with normal kidneys
}
\keyword{internal}
