% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antipsychotics.R
\name{antipsychoticProfile}
\alias{antipsychoticProfile}
\title{Look up an antipsychotic profile}
\usage{
antipsychoticProfile(profile)
}
\arguments{
\item{profile}{a name in \code{antipsychoticProfiles()}}
}
\value{
that profile's entry
}
\description{
Look up an antipsychotic profile
}
