% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs_aprepitant.R
\name{aprepitant}
\alias{aprepitant}
\title{Aprepitant pharmacokinetics}
\usage{
aprepitant(weight, height, age, sex, adjustToFFM = TRUE)
}
\arguments{
\item{weight}{weight in kg}

\item{height}{height in cm}

\item{age}{age in years}

\item{sex}{sex as a string}

\item{adjustToFFM}{evaluate the weight covariates at the pharmacokinetic
weight (TRUE) or at total body weight (FALSE)}
}
\value{
a list in the shape \code{getDrugPK()} expects
}
\description{
One-compartment intravenous disposition of Nijstad 2023, for aprepitant
given intravenously; doses are mg of aprepitant.
}
