% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renalFunction.R
\name{bsaDuBois}
\alias{bsaDuBois}
\title{Du Bois body surface area}
\usage{
bsaDuBois(weight, height)
}
\arguments{
\item{weight}{weight in kg}

\item{height}{height in cm}
}
\value{
body surface area in m^2
}
\description{
Du Bois body surface area
}
\keyword{internal}
