% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs_fosaprepitant.R
\name{fosaprepitant}
\alias{fosaprepitant}
\title{Fosaprepitant pharmacokinetics}
\usage{
fosaprepitant(weight, height, age, sex, adjustToFFM = TRUE)
}
\arguments{
\item{weight}{weight in kg}

\item{height}{height in cm}

\item{age}{age in years}

\item{sex}{sex as a string}

\item{adjustToFFM}{evaluate the weight covariates at the pharmacokinetic
weight (TRUE) or at total body weight (FALSE)}
}
\value{
a list in the shape \code{getDrugPK()} expects
}
\description{
Plasma aprepitant after intravenous fosaprepitant (Nijstad 2023). Doses
are mg of fosaprepitant free acid, converted to aprepitant (0.8699 mg/mg)
on the assumption of instantaneous, complete conversion.
}
