% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanceClosedFormGas.R
\name{gasMixedVenous}
\alias{gasMixedVenous}
\title{Blood-flow-weighted mixed venous tension for one soluble gas}
\usage{
gasMixedVenous(y, body)
}
\arguments{
\item{y}{state vector for one soluble gas, \code{[circ, alv, brain, muscle, fat]}}

\item{body}{geometry from \code{getGasBody()}}
}
\value{
mixed venous tension, percent of one atmosphere
}
\description{
Blood-flow-weighted mixed venous tension for one soluble gas
}
