% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugDefaults.R
\name{getDrugDefaults}
\alias{getDrugDefaults}
\title{Returns the stanpumpR drug defaults table for a single drug}
\usage{
getDrugDefaults(drug)
}
\arguments{
\item{drug}{The drug in question}
}
\value{
A subsetted data.frame from getDrugDefaultsGlobal()
}
\description{
Returns the stanpumpR drug defaults table for a single drug
}
