% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugDefaults.R
\name{getDrugDefaultsGlobal}
\alias{getDrugDefaultsGlobal}
\title{Return the stanpumpR drug defaults table, with expansion applied}
\usage{
getDrugDefaultsGlobal()
}
\description{
Return the stanpumpR drug defaults table, with expansion applied
}
\examples{
getDrugDefaultsGlobal()

}
