% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_absorption.R
\name{ig_absorption_conc}
\alias{ig_absorption_conc}
\title{Plasma concentration with Inverse Gaussian absorption}
\usage{
ig_absorption_conc(t, A, k, mu, lambda, dose = 1, F = 1)
}
\arguments{
\item{t}{numeric vector of times (same units as mu and the exponents)}

\item{A}{coefficients of the unit disposition function}

\item{k}{exponents of the unit disposition function (same length as A)}

\item{mu}{mean absorption time}

\item{lambda}{IG shape parameter (CV^2 of absorption time = mu / lambda)}

\item{dose}{administered dose}

\item{F}{bioavailability}
}
\value{
numeric vector of concentrations, same length as t
}
\description{
Plasma concentration with Inverse Gaussian absorption
}
