% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shiny-ui.R
\name{inputWithChoices}
\alias{inputWithChoices}
\title{Create an input that has choices appended to the right of it. The choices are
technically radio buttons under the hood, but they appear as buttons.}
\usage{
inputWithChoices(tag, choices, inputId = NULL, selected = NULL)
}
\arguments{
\item{tag}{The main input that that should have choices next to it}

\item{choices}{The choices vector, this is passed directly to shiny's \code{radioButtons()}}

\item{inputId}{The input ID to use for the resulting radio buttons. If not provided,
the ID will be a concatenation of the provided input's ID together with "Choice".}

\item{selected}{Which initial choice is selected}
}
\description{
Create an input that has choices appended to the right of it. The choices are
technically radio buttons under the hood, but they appear as buttons.
}
