% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{inputWithInlineLabel}
\alias{inputWithInlineLabel}
\title{Create an input that has its label on its left, and the input takes
the rest of the space}
\usage{
inputWithInlineLabel(tag)
}
\description{
Create an input that has its label on its left, and the input takes
the rest of the space
}
