% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasProperties.R
\name{isGasDrug}
\alias{isGasDrug}
\title{Is this dose-table entry handled by the inhaled-gas engine?}
\usage{
isGasDrug(drug)
}
\arguments{
\item{drug}{one or more drug names}
}
\value{
logical vector
}
\description{
Is this dose-table entry handled by the inhaled-gas engine?
}
