% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{lineTypeSelector}
\alias{lineTypeSelector}
\title{Create a button group for selecting a line type using SVG images instead of text}
\usage{
lineTypeSelector(inputId, label, selected)
}
\description{
Create a button group for selecting a line type using SVG images instead of text
}
