% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs_meloxicam.R
\name{meloxicam}
\alias{meloxicam}
\title{Meloxicam pharmacokinetics (apparent oral)}
\usage{
meloxicam(weight, height, age, sex, adjustToFFM = TRUE, cyp2c9 = "*1/*1")
}
\arguments{
\item{weight}{weight in kg}

\item{height}{height in cm}

\item{age}{age in years}

\item{sex}{sex as a string}

\item{adjustToFFM}{when \code{TRUE}, scale the size-free apparent
parameters (CL/F, Q/F, Vp/F) to fat-free mass; Vc/F follows its own
lean-body-mass covariate either way}

\item{cyp2c9}{CYP2C9 diplotype, one of
\code{"*1/*1", "*1/*2", "*1/*3", "*2/*2", "*2/*3", "*3/*3"}}
}
\value{
a list in the shape \code{getDrugPK()} expects
}
\description{
Meloxicam pharmacokinetics (apparent oral)
}
