% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes.R
\name{oralSaturationForm}
\alias{oralSaturationForm}
\title{The form of an \code{oralSaturation} block: "saturable" unless it says otherwise}
\usage{
oralSaturationForm(saturation)
}
\description{
The form of an \code{oralSaturation} block: "saturable" unless it says otherwise
}
\keyword{internal}
