% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs_oxymorphone.R
\name{oxymorphone}
\alias{oxymorphone}
\title{Oxymorphone pharmacokinetics}
\usage{
oxymorphone(weight, height, age, sex, adjustToFFM = TRUE)
}
\arguments{
\item{weight}{weight in kg}

\item{height}{height in cm (not used)}

\item{age}{age in years (not used)}

\item{sex}{sex as a string (not used)}
}
\value{
a list in the shape \code{getDrugPK()} expects
}
\description{
Oxymorphone pharmacokinetics
}
