% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasProperties.R
\name{potentAgents}
\alias{potentAgents}
\title{The agents that contribute to summed MAC}
\usage{
potentAgents()
}
\value{
character vector of gas names
}
\description{
The agents that contribute to summed MAC
}
