% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoveryStates.R
\name{recoveryFromStates}
\alias{recoveryFromStates}
\title{Time until threshold from a set of effect-site states}
\usage{
recoveryFromStates(set, emerge)
}
\arguments{
\item{set}{a state set from \code{recoveryStateSet()}}

\item{emerge}{the threshold the effect site has to fall to}
}
\value{
minutes, one per row of \code{set$state}; zeros when there is no
threshold to fall to
}
\description{
Time until threshold from a set of effect-site states
}
\keyword{internal}
