% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renalFunction.R
\name{creatinineClearanceCG}
\alias{creatinineClearanceCG}
\title{Cockcroft-Gault creatinine clearance}
\usage{
creatinineClearanceCG(weight, age, sex, scr = assumedCreatinine(sex))
}
\arguments{
\item{weight}{the weight the calling model scales on, kg}

\item{age}{age in years}

\item{sex}{\code{"male"} or \code{"female"}}

\item{scr}{serum creatinine in mg/dL; defaults to the assumed normal value}
}
\value{
creatinine clearance in mL/min
}
\description{
\verb{(140 - age) x weight / (72 x SCr)}, times 0.85 for a woman, in mL/min.
This is the estimator the vancomycin (Thomson 2009), cefazolin (Komatsu
2024) and sugammadex (Kleijn 2011) models were fitted with.
}
\keyword{internal}
