% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanceClosedFormGas.R
\name{gasSystemOxygen}
\alias{gasSystemOxygen}
\title{System matrix and forcing vector for oxygen}
\usage{
gasSystemOxygen(body, Q, VA, Ffgf)
}
\arguments{
\item{body}{output of \code{getGasBody()}}

\item{Q}{total fresh gas flow, L/min}

\item{VA}{alveolar ventilation, L/min}

\item{Ffgf}{fresh-gas oxygen fraction, percent of 1 atm}
}
\value{
a list with \code{A} (2x2) and \code{b} (length 2)
}
\description{
Implements equation (4) for the state ordering \code{c(F_circ, F_alv)}.
}
\keyword{internal}
