% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugDefaults.R
\name{getDrugDefaults}
\alias{getDrugDefaults}
\title{Returns the stanpumpR drug defaults table for a single drug}
\usage{
getDrugDefaults(drug)
}
\description{
Returns the stanpumpR drug defaults table for a single drug
}
\examples{
getDrugDefaults('remifentanil')

}
