% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugAndEventDefaults.R
\name{getDrugDefaultsGlobal}
\alias{getDrugDefaultsGlobal}
\title{Return the stanpumpR drug defaults table, with expansion applied}
\usage{
getDrugDefaultsGlobal()
}
\value{
A data.frame containing concentrations and bolus units, suggested colors and plasma/effect levels for drug effect endpoints
}
\description{
Return the stanpumpR drug defaults table, with expansion applied
}
