% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugAndEventDefaults.R
\name{getDrugDefaultsGlobal}
\alias{getDrugDefaultsGlobal}
\title{Return the stanpumpR drug defaults table}
\usage{
getDrugDefaultsGlobal(expand = TRUE)
}
\arguments{
\item{expand}{If \code{TRUE}, the list in the Units column is expanded. Otherwise,
the Units column contains comma-separated strings.}
}
\value{
A data.frame containing concentrations and bolus units, suggested colors and plasma/effect levels for drug effect endpoints
}
\description{
Return the stanpumpR drug defaults table
}
