% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCpCe.R
\name{simCpCe}
\alias{simCpCe}
\title{Simulate plasma and effect site concentration from time 0 to maximum}
\usage{
simCpCe(dose, events, PK, maximum, plotRecovery)
}
\arguments{
\item{dose}{table of individual doses}

\item{events}{table of events}

\item{PK}{PK parameters from \code{getDrugPK(drug)}}

\item{maximum}{maximum length of simulation in minutes}

\item{plotRecovery}{should the "time until threshold" be calculated?  For
each time point, how long the effect site would take to fall to
\code{PK$endCe} if all delivery stopped at that moment; returned as the
\code{Recovery} column of \code{equiSpace}.  Checked against stopping
delivery in the simulation itself by
\code{tests/testthat/test-recovery-engines.R} (2026-10-05).}
}
\value{
a list of data frames with the output of the a single drug simulation
}
\description{
See \code{vignette("stanpumpR-single-PK", package = "stanpumpR")} for an example
}
