% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCpCe.R
\name{simCpCe}
\alias{simCpCe}
\title{Simulate plasma and effect site concentration from time 0 to maximum}
\usage{
simCpCe(dose, events, PK, maximum, plotRecovery)
}
\arguments{
\item{dose}{table of individual doses}

\item{events}{table of events}

\item{PK}{PK parameters from \code{getDrugPK(drug)}}

\item{maximum}{maximum length of simulation in minutes}

\item{plotRecovery}{(current broken, leave set to FALSE) should recovery parameters be calculated?}
}
\value{
a list of data frames with the output of the a single drug simulation
}
\description{
See \code{vignette("stanpumpR-single-PK", package = "stanpumpR")} for an example
}
