% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDrugsWithCovariates.R
\name{simulateDrugsWithCovariates}
\alias{simulateDrugsWithCovariates}
\title{Process a dose table for multiple drugs using the internal defaults for each drug}
\usage{
simulateDrugsWithCovariates(
  dose,
  events,
  weight,
  height,
  age,
  sex,
  maximum,
  plotRecovery,
  cyp2d6 = CYP2D6_DEFAULT
)
}
\arguments{
\item{dose}{table of individual doses}

\item{events}{table of events}

\item{weight}{weight in kg}

\item{height}{height in cm}

\item{age}{age in years}

\item{sex}{sex as string: "female" or "male"}

\item{maximum}{maximum length of simulation in minutes}

\item{plotRecovery}{should the "time until threshold" be calculated?  See
\code{simCpCe()}.}

\item{cyp2d6}{CYP2D6 metaboliser phenotype, one of \code{CYP2D6_VALUES}.
Only drugs whose model declares it are affected.}
}
\value{
a list of data frames with the output of the a single drug
simulation.  A drug that forms an active metabolite adds its contribution
to the metabolite drug's entry, which is created if that drug was not
dosed directly.
}
\description{
See \code{vignette("stanpumpR-multi-PK", package = "stanpumpR")} for an example
}
