% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDrugsWithCovariates.R
\name{simulateDrugsWithCovariates}
\alias{simulateDrugsWithCovariates}
\title{Process a dose table for multiple drugs using the internal defaults for each drug}
\usage{
simulateDrugsWithCovariates(
  dose,
  events,
  weight,
  height,
  age,
  sex,
  maximum,
  plotRecovery
)
}
\arguments{
\item{dose}{table of individual doses}

\item{events}{table of events}

\item{weight}{weight in kg}

\item{height}{height in cm}

\item{age}{age in years}

\item{sex}{sex as string: "female" or "male"}

\item{maximum}{maximum length of simulation in minutes}

\item{plotRecovery}{(current broken, leave set to FALSE) should recovery parameters be calculated?}
}
\value{
a list of data frames with the output of the a single drug simulation
}
\description{
See \code{vignette("stanpumpR-multi-PK", package = "stanpumpR")} for an example
}
