% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanpumpR-package.R
\docType{package}
\name{stanpumpR-package}
\alias{stanpumpR}
\alias{stanpumpR-package}
\title{stanpumpR: stanpumpR, a PK/PD simulation program}
\description{
stanpumpR is open-source software for pharmacokinetic / pharmacodynamic simulation. It is intended to make pharmacokinetics accessible to facilitate perioperative patient care, teaching, and research. stanpumpR may be freely downloaded and used without restriction for non-commericial purposes.
}
\author{
\strong{Maintainer}: Steven Shafer \email{steven.shafer@stanford.edu} (\href{https://orcid.org/0000-0002-0289-7793}{ORCID})

Authors:
\itemize{
  \item Steven Shafer \email{steven.shafer@stanford.edu} (\href{https://orcid.org/0000-0002-0289-7793}{ORCID})
  \item Dean Attali \email{daattali@gmail.com} (\href{https://orcid.org/0000-0002-5645-3493}{ORCID})
}

}
\keyword{internal}
