% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs_vancomycin.R
\name{vancomycin}
\alias{vancomycin}
\title{Vancomycin pharmacokinetics}
\usage{
vancomycin(weight, height, age, sex, adjustToFFM = TRUE)
}
\arguments{
\item{weight}{weight in kg}

\item{height}{height in cm}

\item{age}{age in years}

\item{sex}{sex as a string}

\item{adjustToFFM}{scale the size-free parameters to fat-free mass and
estimate creatinine clearance on the pharmacokinetic weight}
}
\value{
a list in the shape \code{getDrugPK()} expects
}
\description{
Vancomycin pharmacokinetics
}
